API

Pioran.SHO — Type

SHO(σ, ω₀, Q)

Construct a simple harmonic oscillator covariance function with parameters σ, ω₀, Q. Where σ is the amplitude, ω₀ is the angular frequency and Q is the quality factor.

Pioran.Celerite — Type

Celerite(a, b, c, d)

Construct a celerite covariance function with parameters a, b, c, d.

Pioran.Exp — Type

Exp(σ,α)

Construct a exponential covariance function with parameters σ, α. Where σ is the amplitude and α is the decay rate.

Pioran.posterior — Function

posterior(f::ScalableGP, y::AbstractVecOrMat{<:Real})

Compute the posterior GP given the GP f and the data y.

Pioran.approx — Function

approx(psdmodel, f0, fM, ncomponents=20, var=1.0; basis_function="SHO")

Approximate the PSD with a sum of SHO functions

psd_model: the model for the power spectral density
f0: the lowest frequency
fM: the highest frequency
n_components: the number of components to use
var: the variance of the process
basis_function: the basis function to use

Pioran.log_likelihood — Function

Compute the log likelihood of the GP at the points τ given the data y and time t.

Pioran.plot_mean_approx — Function

Plot the frequency-averaged residuals and ratios

Pioran.plot_quantiles_approx — Function

plot_quantiles_approx(f, f_min, f_max, residuals, ratios; path="")

Plot the quantiles of the residuals and ratios (with respect to the approximated PSD) of the PSD

Pioran.plot_boxplot_psd_approx — Function

Plot the boxplot of the residuals and ratios for the PSD approximation

Pioran.run_diagnostics — Function

Run the diagnostics for the PSD approximation

Pioran.extract_subset — Function

extractsubset(rng, prefix, t, y, yerr; nperc=0.03, take_log=true)

Extract a subset of the data for the analysis and return initial guesses for the mean and variance.

Parameters

rng : Random.MersenneTwister Random number generator. prefix : String Prefix for the output files. t : Array{Float64,1} Time array. y : Array{Float64,1} Time series array. yerr : Array{Float64,1} Time series error array. nperc : Float64 Percentage of the time series to extract. takelog : Bool If true, log transform the time series for the estimation of the mean and variance.

Returns

tsubset : Array{Float64,1} Time array of the subset. ysubset : Array{Float64,1} Time series array of the subset. yerr_subset : Array{Float64,1} Time series error array of the subset. x̄ : Float64 Mean of the normal distribution for μ. va : Float64 Variance of the normal distribution for μ.

<!– @autodocs Modules = [Pioran] Order = [:function, :type] –>