API Reference

Power spectral densities

Models and functions

Pioran.DoubleBendingPowerLaw — Type

 DoubleBendingPowerLaw(α₁, f₁, α₂, f₂, α₃)

Double bending power law model for the power spectral density

α₁: the first power law index

f₁: the first bend frequency

α₂: the second power law index

f₂: the second bend frequency

α₃: the third power law index

\[\mathcal{P}(f) = \frac{(f/f₁)^{-α₁}}{1 + (f / f₁)^{α₂ - α₁}}\frac{1}{1 + (f / f₂)^{α₃ - α₂}}\]

Pioran.DoubleBendingPowerLaw_Bis — Type

 DoubleBendingPowerLaw_Bis(α₀, f₁, Δα₁, Δf, Δα₂)

Double bending power law model for the power spectral density

α₀: the first power law index

f₁: the first bend frequency

Δα₁: the first difference in power law index

Δf: scale for the second bend frequency, f₂ = f₁ * Δf

Δα₂: the second difference in power law index

\[\mathcal{P}(f) = \frac{(f/f₁)^{-α_0}}{1 + (f / f₁)^{α_0+\Delta α₁}}\frac{1}{1 + (f / f₁ \Delta f)^{\Delta α₁ + \Delta α₂}}\]

Pioran.SingleBendingPowerLaw — Type

 SingleBendingPowerLaw(α₁, f₁, α₂)

Single bending power law model for the power spectral density

α₁: the first power law index

f₁: the first bend frequency

α₂: the second power law index

\[\mathcal{P}(f) = \frac{(f/f₁)^{-α₁}}{1 + (f / f₁)^{α₂ - α₁}}\]

Pioran.approx — Function

 approx(psd_model, f0, fM, n_components=20, var=1.0; basis_function="SHO")

Approximate the PSD with a sum of basis functions to form a covariance function

Arguments

psd_model::PowerSpectralDensity: model of the PSD

f0::Real: the lowest frequency

fM::Real: the highest frequency

n_components::Integer=20: the number of basis functions to use

var::Real=1.0: the variance of the process, integral of the PSD

basis_function::String="SHO": the basis function to use, either "SHO" or "DRWCelerite"

Return

covariance::SumOfSemiSeparable: the covariance function

Example

using Pioran
𝓟 = SingleBendingPowerLaw(1.0, 1.0, 2.0)
𝓡 = approx(𝓟, 1e-4, 1e-1, 30, 2.31,basis_function="SHO")

Helper functions

Pioran.approximated_psd — Method

 approximated_psd(f, psd_model, f0, fM; n_components=20, var=1.0, basis_function="SHO")

Return the approximated PSD. This is essentially to check that the model and the approximation are consistent.

Arguments

f::AbstractVector{<:Real}: the frequencies at which to calculate the PSD

psd_model::PowerSpectralDensity: model of the PSD

f0::Real: the lowest frequency

fM::Real: the highest frequency

n_components::Integer=20: the number of basis functions to use

var::Real=1.0: the variance of the process, integral of the PSD

basis_function::String="SHO": the basis function to use, either "SHO" or "DRWCelerite"

individual::Bool=false: return the individual components

Pioran.build_approx — Method

 build_approx(J, f0, fM; basis_function="SHO")

Prepare the approximation of a PSD

Arguments

J::Integer: the number of basis functions

f0::Real: the lowest frequency

fM::Real: the highest frequency

basis_function::String="SHO": the basis function to use, either "SHO" or "DRWCelerite"

Return

spectral_points::Vector{Real}: the spectral points

spectral_matrix::Matrix{Real}: the spectral matrix

Pioran.calculate — Method

calculate(f, psd::PowerSpectralDensity)

Calculate the power spectral density at frequency f

Pioran.get_approx_coefficients — Method

 get_approx_coefficients(psd_model, f0, fM; n_components=20, basis_function="SHO")

Get the coefficients of the approximated PSD

Arguments

psd_model::PowerSpectralDensity: model of the PSD

f0::Real: the lowest frequency

fM::Real: the highest frequency

n_components::Integer=20: the number of basis functions to use

basis_function::String="SHO": the basis function to use, either "SHO" or "DRWCelerite"

Return

amplitudes::Vector{Real}: the amplitudes of the basis functions

Pioran.get_normalised_psd — Method

 get_normalised_psd(psd_model::PowerSpectralDensity, spectral_points::AbstractVector{<:Real})

Get the PSD normalised at the lowest frequency

Pioran.psd_decomp — Method

 psd_decomp(psd_normalised, spectral_matrix)

Get amplitudes of the basis functions by solving the linear system of the approximation

Covariance functions

Models and functions

Pioran.Exp — Type

Exp(A, α)

Exponential covariance Function

A: the amplitude of the covariance function

α: the decay rate of the covariance function

\[k(τ) = A \exp(-α τ)\]

Example

Exp(1.0, 0.25)

Pioran.SHO — Type

 SHO(A, ω₀, Q)

Simple Harmonic Oscillator covariance Function

A: the amplitude of the covariance function

ω₀: the angular frequency of the simple harmonic oscillator

Q: the quality factor of the simple harmonic oscillator

\[k(τ) = A \exp(-ω₀ τ / Q / 2) \left\{\begin{matrix} 2(1 + ω₀ τ) & Q = 1/2 \\ \cos(η ω₀ τ) + \frac{\sin(η ω₀ τ)}{2η Q} & Q < 1/2 \\ \cosh(η ω₀ τ) + \frac{\sinh(η ω₀ τ)}{2η Q} & Q \geq 1/2 \end{matrix}\right.\\ η = \sqrt{\left|1 - \frac{1}{4 Q^2}\right|}\]

See Foreman-Mackey et al. (2017) for more details.

Pioran.Celerite — Type

 Celerite(a, b, c, d)

Celerite covariance Function

a: the amplitude of the first term

b: the amplitude of the second term

c: the decay rate of the covariance function

d: the period of the covariance function

\[k(τ) = \exp(-c τ) (a \cos(d τ) + b \sin(d τ))\]

See Foreman-Mackey et al. (2017) for more details.

Helper functions

Base.:+ — Method

 +(::SemiSeparable, ::SemiSeparable)

Sum of semi-separable covariance functions

Pioran.celerite_coefs — Method

 celerite_coefs(covariance)

Get the celerite coefficients

Gaussian processes

Pioran.ScalableGP — Type

A scalable Gaussian process has a covariance function formed of semi-separable kernels

Example

using Pioran
𝓟 = SingleBendingPowerLaw(1.0, 1.0, 2.0)
𝓡 = approx(𝓟, 1e-4, 1e-1, 30, 2.31,basis_function="SHO")
μ = 1.2

f = ScalableGP(𝓡) # zero-mean GP
f = ScalableGP(μ, 𝓡) # with mean μ

See Foreman-Mackey et al. (2017) for more details.

Base.rand — Function

rand(rng::AbstractRNG, fp::PosteriorGP, N::Int=1)
rand(rng::AbstractRNG, fp::PosteriorGP, τ::AbstractVecOrMat{<:Real}, N::Int=1)
rand(fp::PosteriorGP, N::Int=1)

Sample N realisations from the posterior GP fp at the points τ.

Base.rand — Method

rand(rng::AbstractRNG, f::ScalableGP)
rand(rng::AbstractRNG, f::ScalableGP, t::AbstractVecOrMat{<:Real})
rand(f::ScalableGP)
rand(f::ScalableGP, t::AbstractVecOrMat{<:Real})

Draw a realisation from the GP f at the points t.

Distributions.logpdf — Method

logpdf(f::ScalableGP, Y::AbstractVecOrMat{<:Real})

Compute the log-likelihood of the data Y given the GP f.

Pioran._predict_cov — Method

_predict_cov(fp::PosteriorGP, τ::AbstractVecOrMat{<:Real})
Compute the posterior covariance of the GP at the points τ.

Pioran._predict_mean — Method

_predict_mean(fp::PosteriorGP, τ::AbstractVecOrMat{<:Real})

Compute the Posterior mean of the GP at the points τ.

Pioran.posterior — Method

posterior(f::ScalableGP, y::AbstractVecOrMat{<:Real})

Compute the posterior Gaussian process fp given the GP f and the data y.

Statistics.cov — Method

cov(fp::PosteriorGP, τ::AbstractVecOrMat{<:Real})
cov(fp::PosteriorGP)

Compute the covariance of the posterior GP at the points τ.

Statistics.mean — Method

mean(fp::PosteriorGP, τ::AbstractVecOrMat{<:Real})
mean(fp::PosteriorGP)

Compute the mean of the posterior GP at the points τ.

Statistics.std — Method

std(fp::PosteriorGP, τ::AbstractVecOrMat{<:Real})
std(fp::PosteriorGP)

Compute the standard deviation of the posterior GP at the points τ.

Solvers

Celerite solver

Pioran.init_semi_separable! — Method

 init_semi_separable(a, b, c, d, τ, σ2)

Initialise the matrices and vectors needed for the celerite algorithm. U,V are the rank-R matrices, D is the diagonal matrix and ϕ is the matrix of the exponential terms.

See Foreman-Mackey et al. (2017) for more details.

Pioran.log_likelihood — Method

log_likelihood(cov, τ, y, σ2)

Compute the log-likelihood of a semi-separable covariance function using the celerite algorithm.

Arguments

cov::SumOfSemiSeparable or cov::CARMA or cov::SemiSeparable: the covariance function

τ::Vector: the time points

y::Vector: the data

σ2::Vector: the measurement variances

Pioran.logl — Method

logl(a, b, c, d, τ, y, σ2)

Compute the log-likelihood of a GP with a semi-separable covariance function using the celerite algorithm.

Arguments

a::Vector

b::Vector

c::Vector

d::Vector

τ::Vector: the time points

y::Vector: the data

σ2::Vector: the measurement variances

See Foreman-Mackey et al. (2017) for more details.

Pioran.predict — Method

predict(cov, τ, t, y, σ2)

Compute the posterior mean of the GP at the points τ given the data y and time t.

Arguments

cov::SumOfSemiSeparable or cov::CARMA or cov::SemiSeparable: the covariance function

τ::Vector: the time points

t::Vector: the data time points

y::Vector: the data

σ2::Vector: the measurement variances

Pioran.simulate — Method

simulate(rng, cov, τ, σ2)
simulate(cov, τ, σ2)

Draw a realisation from the GP with the covariance function cov at the points τ with the variances σ2.

Arguments

rng::AbstractRNG: the random number generator

cov::SumOfSemiSeparable or cov::CARMA or cov::SemiSeparable: the covariance function

τ::Vector: the time points

σ2::Vector: the measurement variances

Pioran.solve_prec! — Method

solve_prec!(z, y, U, W, D, ϕ)

Forward and backward substitution of the celerite algorithm.

See Foreman-Mackey et al. (2017) for more details.

Direct solver

Pioran.log_likelihood_direct — Method

log_likelihood_direct(cov::KernelFunctions.SimpleKernel, t::Vector, y::Vector, σ²::Vector)

Compute the log-likelihood of the data Y given the GP f with the direct solver.

Pioran.predict_cov — Method

predict_direct(cov::KernelFunctions.SimpleKernel, τ::AbstractVector, t::AbstractVector, σ²::AbstractVector)

Compute the posterior covariance of the GP at the points τ given the times t and the noise variance σ².

Pioran.predict_direct — Function

predict_direct(cov::KernelFunctions.SimpleKernel, τ::AbstractVector, t::AbstractVector, y::AbstractVector, σ²::AbstractVector, with_covariance::Bool=false)

Compute the posterior mean of the GP at the points τ given the data (t, y) and the noise variance σ².

Plotting

Diagnostics

Pioran.run_diagnostics — Method

run_diagnostics(prior_samples, variance_samples, f0, fM, model, f_min, f_max; path="", basis_function="SHO", n_components=20)

Run the prior predictive checks for the model and the approximation of the PSD

Arguments

prior_samples::Array{Float64, 2} : Model samples from the prior distribution

variance_samples::Array{Float64, 1} : The variance samples

f0::Float64 : The minimum frequency for the approximation of the PSD

fM::Float64 : The maximum frequency for the approximation of the PSD

model::Function : The model

f_min::Float64 : The minimum frequency of the observed data

f_max::Float64 : The maximum frequency of the observed data

path::String="" : The path to save the plots

basis_function::String="SHO" : The basis function for the approximation of the PSD

n_components::Int=20 : The number of components to use for the approximation of the PSD

Pioran.run_posterior_predict_checks — Method

run_posterior_predict_checks(samples, paramnames, t, y, yerr, f0, fM, model, with_log_transform; plots="all", n_samples=200, path="", basis_function="SHO", n_frequencies=1000, plot_f_P=false, n_components=20)

Run the posterior predictive checks for the model and the approximation of the PSD

Arguments

samples::Array{Float64, 2} : The samples from the posterior distribution

paramnames::Array{String, 1} : The names of the parameters

t::Array{Float64, 1} : The time series

y::Array{Float64, 1} : The values of the time series

yerr::Array{Float64, 1} : The errors of the time series

f0::Float64 : The minimum frequency for the approximation of the PSD

fM::Float64 : The maximum frequency for the approximation of the PSD

model::Function : The model

with_log_transform::Bool : If true, the flux is log-transformed

plots::String or Array{String, 1} : The type of plots to make. It can be "all", "psd", "lsp", "timeseries" or a combination of them in an array

n_samples::Int=200 : The number of samples to draw from the posterior predictive distribution

path::String="all" : The path to save the plots

basis_function::String="SHO" : The basis function for the approximation of the PSD

n_frequencies::Int=1000 : The number of frequencies to use for the approximation of the PSD

plot_f_P::Bool=false : If true, the plots are made in terms of f * PSD

n_components::Int=20 : The number of components to use for the approximation of the PSD

Pioran.sample_approx_model — Method

sample_approx_model(samples, variance_samples, f0, fM, model; n_frequencies=1_000, basis_function="SHO", n_components=20)

Check the approximation of the PSD by computing the residuals and the ratios of the PSD and the approximated PSD

Arguments

samples::Array{Float64, n} : The model samples

variance_samples::Array{Float64, 1} : The variance samples

f0::Float64 : The minimum frequency for the approximation of the PSD

fM::Float64 : The maximum frequency for the approximation of the PSD

model::Function : The model

n_frequencies::Int=1_000 : The number of frequencies to use for the approximation of the PSD

basis_function::String="SHO" : The basis function for the approximation of the PSD

n_components::Int=20 : The number of components to use for the approximation of the PSD

Return

psd::Array{Float64, 2} : The PSD

psd_approx::Array{Float64, 2} : The approximated PSD

residuals::Array{Float64, 2} : The residuals (psd-approx_psd)

ratios::Array{Float64, 2} : The ratios (approx_psd/psd)

Individual plotting functions

Pioran.get_ppc_timeseries — Method

get_ppc_timeseries(samples_𝓟, samples_variance, samples_ν, samples_μ, t, y, yerr,f0,fM, model,with_log_transform;samples_c=missing, n_samples=1000)

Get the random posterior predictive time series from the model and samples

Pioran.plot_boxplot_psd_approx — Method

plot_boxplot_psd_approx(residuals, ratios; path="")

Plot the boxplot of the residuals and ratios for the PSD approximation

Pioran.plot_lsp_ppc — Method

plot_lsp_ppc(samples_𝓟, samples_variance, samples_ν, samples_μ, t, y, yerr, f0, fM, model; plot_f_P=false, n_frequencies=1000, n_samples=1000, n_components=20, bin_fact=10, path="", basis_function="SHO")

Plot the posterior predictive Lomb-Scargle periodogram of random time series from the model and samples

Arguments

samples_𝓟::Array{Float64, n} : The samples of the model parameters

samples_variance::Array{Float64, 1} : The variance samples

samples_ν::Array{Float64, 1} : The ν samples

samples_μ::Array{Float64, 1} : The μ samples

t::Array{Float64, 1} : The time series

y::Array{Float64, 1} : The values of the time series

yerr::Array{Float64, 1} : The errors of the time series

f0::Float64 : The minimum frequency for the approximation of the PSD

fM::Float64 : The maximum frequency for the approximation of the PSD

model::Function : The model

plot_f_P::Bool=false : If true, the plot is made in terms of f * PSD

n_frequencies::Int=1000 : The number of frequencies to use for the approximation of the PSD

n_samples::Int=1000 : The number of samples to draw from the posterior predictive distribution

n_components::Int=20 : The number of components to use for the approximation of the PSD

bin_fact::Int=10 : The binning factor for the LSP

path::String="" : The path to save the plots

basis_function::String="SHO" : The basis function for the approximation of the PSD

Pioran.plot_mean_approx — Method

plot_mean_approx(f, residuals, ratios; path="")

Plot the frequency-averaged residuals and ratios

Pioran.plot_ppc_timeseries — Method

plot_ppc_timeseries(samples_𝓟, samples_variance, samples_ν, samples_μ, t, y, yerr, f0, fM, model, with_log_transform; t_pred=nothing, samples_c=nothing, n_samples=100, n_components=20, basis_function="SHO", path="")

Plot the posterior predictive time series and the residuals

Arguments

samples_𝓟::Matrix{Float64} : The samples of the model parameters

samples_variance::Vector{Float64} : The variance samples

samples_ν::Vector{Float64} : The noise level samples

samples_μ::Vector{Float64} : The mean samples

t::Vector{Float64} : The time series

y::Vector{Float64} : The values of the time series

yerr::Vector{Float64} : The errors of the time series

f0::Float64 : The minimum frequency for the approximation of the PSD

fM::Float64 : The maximum frequency for the approximation of the PSD

model::PowerSpectralDensity : The model

with_log_transform::Bool : If true, the flux was log-transformed for the inference

t_pred::Vector{Float64}=nothing : The prediction times

samples_c::Vector{Float64}=nothing : The samples of the constant (if the flux was log-transformed)

n_samples::Int=100 : The number of samples to draw from the posterior predictive distribution

n_components::Int=20 : The number of components to use for the approximation of the PSD

basis_function::String="SHO" : The basis function for the approximation of the PSD

Pioran.plot_psd_ppc — Method

plot_psd_ppc(samples_𝓟, samples_variance, samples_ν, t, y, yerr, f0, fM, model; plot_f_P=false, n_frequencies=1000, path="", n_components=20, basis_function="SHO", with_log_transform=false)

Plot the posterior predictive power spectral density and the noise levels

Arguments

samples_𝓟::Array{Float64, n} : The samples of the model parameters

samples_variance::Array{Float64, 1} : The variance samples

samples_ν::Array{Float64, 1} : The ν samples

t::Array{Float64, 1} : The time series

y::Array{Float64, 1} : The values of the time series

yerr::Array{Float64, 1} : The errors of the time series

f0::Float64 : The minimum frequency for the approximation of the PSD

fM::Float64 : The maximum frequency for the approximation of the PSD

model::Function : The model

plot_f_P::Bool=false : If true, the plot is made in terms of f * PSD

n_frequencies::Int=1000 : The number of frequencies to use for the approximation of the PSD

n_components::Int=20 : The number of components to use for the approximation of the PSD

basis_function::String="SHO" : The basis function for the approximation of the PSD

path::String="" : The path to save the plots

Pioran.plot_quantiles_approx — Method

plot_quantiles_approx(f, f_min, f_max, residuals, ratios; path="")

Plot the quantiles of the residuals and ratios (with respect to the approximated PSD) of the PSD

Pioran.plot_residuals_diagnostics — Method

plot_residuals_diagnostics(t, mean_res, res_quantiles; confidence_intervals=[95, 99], path="")

Plot the residuals and the autocorrelation function of the residuals

Arguments

t::Vector{Float64} : The time series

mean_res::Vector{Float64} : The mean of the residuals

res_quantiles::Array{Float64, 2} : The quantiles of the residuals

confidence_intervals::Array{Int, 1} : The confidence intervals

path::String="" : The path to save the plot

Pioran.plot_simu_ppc_timeseries — Method

plot_simu_ppc_timeseries(t_pred, ts_quantiles, t, y, yerr; path="")

Plot the posterior predictive simulated time series

Arguments

t_pred::Vector{Float64} : The prediction times

ts_quantiles::Array{Float64, 2} : The quantiles of the posterior predictive time series

t::Vector{Float64} : The time series

y::Vector{Float64} : The values of the time series

yerr::Vector{Float64} : The errors of the time series

path::String="" : The path to save the plot

Utilities

Pioran.extract_subset — Method

extract_subset(rng, prefix, t, y, yerr; n_perc=0.03, take_log=true, suffix="")
extract_subset(seed, prefix, t, y, yerr; n_perc=0.03, take_log=true)

Extract a subset of the data for the analysis and return initial guesses for the mean and variance. Either a random number generator or a seed can be provided.

Arguments

seed::Int64 : Seed for the random number generator.

rng::AbstractRNG : Random number generator.

prefix::String : Prefix for the output files.

t::Array{Float64,1} : Time array.

y::Array{Float64,1} : Time series array.

yerr::Array{Float64,1} : Time series error array.

n_perc::Float64 : Percentage of the time series to extract.

take_log::Bool : If true, log transform the time series for the estimation of the mean and variance.

suffix::String : Suffix for the output files.

Return

t_subset::Array{Float64,1} : Time array of the subset.

y_subset::Array{Float64,1} : Time series array of the subset.

yerr_subset::Array{Float64,1} : Time series error array of the subset.

x̄::Float64 : Mean of the normal distribution for μ.

va::Float64 : Variance of the normal distribution for μ.

Pioran.separate_samples — Method

separate_samples(samples, paramnames, with_log_transform::Bool)

Separate the samples into the parameters of the model and the parameters of the power spectral density.

Prior distributions

Pioran.ThreeUniformDependent — Type

ThreeUniformDependent(a, b, c, ϵ)
ThreeUniformDependent(a, b, c) (constructor with default ϵ = 1e-10)

Multivariate distribution to model three random variables where the first one x1 is given by U[a,b] and the second one x2 is given by U[x1,c] and the third one x3 is given by U[x2,c]. where a<b<c.

a: lower bound of the first distribution

b: upper bound of the first distribution

c: upper bound of the second and third distribution

ϵ: small value to make sure that the lower and upper bounds of each distribution are different

This means that the lower bound of the second distribution is dependent on the value of the first distribution and so on... This is implemented to overcome the limitations of the current Turing's implementation for dependent priors with dynamic support. See the following issues for more details: [1],[2],[3].

Pioran.TwoLogUniformDependent — Type

TwoLogUniformDependent(a, b, ϵ) 
TwoLogUniformDependent(a, b) (constructor with default ϵ = 1e-10

Multivariate distribution to model three random variables where the first one x1 is given by log-U[a,b] and the second one x2 is given by log-U[x1,b].

a: lower bound of the first distribution

b: upper bound of the first distribution

ϵ: small value to make sure that the lower and upper bounds of each distribution are different

This means that the lower bound of the second distribution is dependent on the value of the first distribution. This is implemented to overcome the limitations of the current Turing's implementation for dependent priors with dynamic support. See the following issues for more details: [1],[2],[3].

Pioran.TwoUniformDependent — Type

TwoUniformDependent(a, b, c, ϵ)
TwoUniformDependent(a, b, c) (constructor with default ϵ = 1e-10)

Multivariate distribution to model two random variables where the first one is given by U[a,b] and the second one is given by U[x,c], where x is a random variable sampled from the first distribution.

a: lower bound of the first distribution

b: upper bound of the first distribution

c: upper bound of the second distribution

ϵ: small value to make sure that the lower and upper bounds of each distribution are different

This means that the lower bound of the second distribution is dependent on the value of the first distribution.This is implemented to overcome the limitations of the current Turing's implementation for dependent priors with dynamic support. See the following issues for more details: [1],[2],[3].

Example

julia> using Pioran, Distributions
julia> d = TwoUniformDependent(0, 1, 2)
TwoUniformDependent(0.0, 1.0, 2.0)

julia> rand(d)
2-element Array{Float64,1}:
 0.123
 1.234

Bijectors.bijector — Method

Bijectors.bijector(d::TwoUniformDependent)

Create a bijector for the TwoUniformDependent distribution. This is used to sample from the distribution using the Bijectors package. Adapted from the following issues [1],[2],[3].

CARMA

Pioran.CARMA — Type

CARMA(p, q, rα, β, σ²)

Continuous-time AutoRegressive Moving Average (CARMA) model for the power spectral density

p: the order of the autoregressive polynomial

q: the order of the moving average polynomial

rα: roots of the autoregressive polynomial length p+1

β: the moving average coefficients length q+1

σ²: the variance of the process

The power spectral density of the CARMA model is given by:

\[\mathcal{P}(f) = \sigma^2 \left|\dfrac{\sum\limits_{k=0}^q \beta_k \left(2\pi\mathrm{i}f\right)^k }{\sum\limits_{l=0}^p \alpha_l \left(2\pi\mathrm{i}f\right)^l}\right|^2\]

Pioran.quad2roots — Method

quad2roots(quad)

Convert the coefficients of a quadratic polynomial to its roots.

Arguments

quad::Vector{Real}: Coefficients of the quadratic polynomial.

Returns

r::Vector{Complex}: Roots of the polynomial.

Pioran.roots2coeffs — Method

roots2coeffs(r)

Convert the roots of a polynomial to its coefficients.

Arguments

r::Vector{Complex}: Roots of the polynomial.

Returns

c::Vector{Complex}: Coefficients of the polynomial.